调整pom.xml GAV & 解决插入id自增问题

+++ datafactory-core/src/main/java/com/coderman/tianhua/datafactory/core/mapper/DataSourceMapper.java
@@ -24,7 +24,7 @@ public interface DataSourceMapper{
 	 * @return int
 	 * @throws
 	 */
-	public int insert(DataSourceEntity dataSourceEntity);
+	public long insert(DataSourceEntity dataSourceEntity);
 
 	/**
 	 *
